Do not include the download_url and pypi_name in the json.
This was only included when an old-style product zip was added, but future hotfixes will not have this, because this does not work on Plone 5.2+.
The json was not really used anyway I think, at least not by our own code, and it was not publicly advertised as far as I know.

This brings the json size down from 200 to 100 kB.

Note that there is still duplication:
if a hotfix is applicable for ten Plone versions, its info will be there ten times.
Could be improved, but that is nice to have for later.